# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ai Pop Cute Support Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: getFeed
      tags: [posts]
      summary: Get home feed
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Feed posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post by ID
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Post not found
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Post updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Post deleted
  /posts/{postId}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
  /posts/{postId}/unlike:
    post:
      operationId: unlikePost
      tags: [posts]
      summary: Unlike a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unliked
  /posts/{postId}/comments:
    get:
      operationId: getComments
      tags: [posts]
      summary: Get post comments
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentListResponse"
    post:
      operationId: createComment
      tags: [posts]
      summary: Add a comment
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /posts/{postId}/comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [posts]
      summary: Delete a comment
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Comment deleted
  /posts/{postId}/pin:
    post:
      operationId: pinPost
      tags: [posts]
      summary: Pin a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post pinned
  /posts/{postId}/unpin:
    post:
      operationId: unpinPost
      tags: [posts]
      summary: Unpin a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post unpinned
  /posts/{postId}/share:
    post:
      operationId: sharePost
      tags: [posts]
      summary: Share a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post shared
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, numericId, name, username, isVerified, role, followersCount, followingCount, postsCount, likesCount, createdAt]
      properties:
        id:
          type: integer
        numericId:
          type: integer
        name:
          type: string
        username:
          type: string
        bio:
          type: ["string", "null"]
        alternateBio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        role:
          type: string
          enum: [user, creator]
        isPrivate:
          type: boolean
        mixChannelId:
          type: ["string", "null"]
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        likesCount:
          type: integer
        totalVotes:
          type: integer
        profileViews:
          type: integer
        isFollowing:
          type: ["boolean", "null"]
        isFollowRequestSent:
          type: ["boolean", "null"]
        isBlocked:
          type: ["boolean", "null"]
        createdAt:
          type: string
          format: date-time
...
    Post:
      type: object
      required: [id, authorId, visibility, allowComments, allowDownloads, isPinned, viewsCount, likesCount, commentsCount, sharesCount, createdAt]
      properties:
        id:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/UserProfile"
        caption:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        mediaType:
          type: ["string", "null"]
          enum: ["image", "video", null]
        filterName:
          type: ["string", "null"]
        visibility:
          type: string
          enum: [public, followers, private]
        allowComments:
          type: boolean
        allowDownloads:
          type: boolean
        isPinned:
          type: boolean
        viewsCount:
          type: integer
        likesCount:
          type: integer
        commentsCount:
          type: integer
        sharesCount:
          type: integer
        downloadsCount:
          type: integer
        isLiked:
          type: ["boolean", "null"]
        isSaved:
          type: ["boolean", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PostInput:
      type: object
      properties:
        caption:
          type: ["string", "null"]
        mediaUrl:
          type: ["string", "null"]
        mediaType:
          type: ["string", "null"]
        filterName:
          type: ["string", "null"]
        visibility:
          type: string
          enum: [public, followers, private]
        allowComments:
          type: boolean
        allowDownloads:
          type: boolean
        isPinned:
          type: boolean
    PostUpdate:
      type: object
      properties:
        caption:
          type: ["string", "null"]
        visibility:
          type: string
          enum: [public, followers, private]
        allowComments:
          type: boolean
        allowDownloads:
          type: boolean
        isPinned:
          type: boolean
    PostListResponse:
      type: object
      required: [posts, total, page]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        hasMore:
          type: boolean
    LikeResponse:
      type: object
      required: [liked, likesCount]
      properties:
        liked:
          type: boolean
        likesCount:
          type: integer
    Comment:
      type: object
      required: [id, postId, authorId, content, createdAt]
      properties:
        id:
          type: integer
        postId:
          type: integer
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/UserProfile"
        parentId:
          type: ["integer", "null"]
        content:
          type: string
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
          format: date-time
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        parentId:
          type: ["integer", "null"]
    CommentListResponse:
      type: object
      required: [comments, total, page]
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        total:
          type: integer
        page:
          type: integer

    # ── Vote Schemas ─────────────────────────────────────────────────────────
...